requirements_web.txt
# Web app requirements (lightweight for inference only)
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.0.0
werkzeug>=2.3.0

# Optional: for ONNX inference
onnxruntime>=1.14.0
